+++ Runtime/Api/QosManager.cs
@@ -74,6 +74,10 @@ namespace AccelByte.Api
             callback.TryOk(latencies);
         }
 
+        /// <summary>
+        /// Get all server latencies available
+        /// </summary>
+        /// <param name="callback">Returns a result via callback when completed</param>
         public void GetAllServerLatencies( ResultCallback<Dictionary<string, int>> callback )
         {
             coroutineRunner.Run(GetAllServerLatenciesAsync(callback));
@@ -95,6 +99,10 @@ namespace AccelByte.Api
             yield return CalculateServerLatencies(getQosServersResult.Value.servers, callback);
         }
 
+        /// <summary>
+        /// Get server latencies for active QoS server in the namespace.
+        /// </summary>
+        /// <param name="callback">Returns a result via callback when completed</param>
         public void GetServerLatencies(ResultCallback<Dictionary<string, int>> callback)
         {
             coroutineRunner.Run(GetServerLatenciesAsync(callback));